Add `NSErrorException` class
Kotlin/Native can't throw an `NSError::class`. So, we make an `NSErrorException` wrapper that is `Throwable` that contains the underlying `NSError`.

This helps us propagate `NSError`s to calling code. Callers an inspect the underlying error, instead of being stuck with just a `KotlinException`.

In your Swift code, use this `Error` extension to make it easy to inspect the underlying error:

```
extension Error {
    var wrappedNSError: NSError? {
        let kotlinException = (self as NSError).userInfo["KotlinException"] as? KotlinException
        let nsErrorException = kotlinException as? NSErrorException
        return nsErrorException?.nsError as? NSError
    }
}
```

Example usage:

```
do {
    try kotlinMethodThatThrowsNSErrorException()
} catch {
    if let wrappedError = error.wrappedNSError {
        print("wrappedError domain: \(wrappedError.domain)")
        print("wrappedError code: \(wrappedError.code)")
        print("wrappedError userInfo: \(wrappedError.userInfo)")
    }
}
```