Only allow proper identifiers as strings keys
The old code checked if key and text are identical, which made no sense
as there is no such limitation. On the other hand it allowed keys like
"test-key" which are invalid on Android.

The new code checks if the identifier starts with an ASCII letter or
underscore and only contains ASCII lettes, digits or underscores.